sample: exponential sampling code
We will simulate a memoryless process that gives each allocated byte
the same probability of being sampled.  Rather than flipping
a (biased) coin for every byte, we instead generate the number of
consecutive "don't sample" decisions with an Exponential distribution.

The current hard-coded value samples, on average, once every 32MB.

There is a new test for the PRNG in sample.c; build it by with
`BUILD_TESTS=1 ./make.sh`.  The output should be a stream of
values sampled uniformly from `(0, 1]`.