tracked_alloc: tracking allocator
In order to generate tracepoint events for `free` and `realloc` calls
on allocations that were previously sampled, we need a to fulfill
these requests with a backup tracking allocator that can quickly
identify whether a given address was tracked and returned by that
allocator.  It's important that we quickly determine when allocations
are not tracked, in order to minimise the overhead in most calls to
`free` and `realloc`.

Our sampling logic only tracks a small fraction of the full heap
footprint (hardcoded to `1 / (32 * 1024 * 1024)`, i.e., less than one
in a million), so we can be wasteful when it comes to the few
allocations we do track.

The tracking allocator directly uses mmap/munmap, and only returns
addresses that are aligned to 1GB: this gives a simple arithmetic test
to quickly determine that most addresses are definitely not tracked,
as well as a simple bucketing scheme to test whether a given aligned
address came from the tracking allocator.

In the future, we will want to offset addresses way from 1GB
boundaries, e.g., to insert guard pages before or after allocations.
A reasonable compromise might be to allow offsets up to 8 KB away from
the aligned base: that's still a tiny fraction of the 1 GB range.