poireau.py: track when we reach a new max allocation footprint
When the first argument is `--track-high-water-mark`.  In that case,
the second argument may be an integer, which is the minimum size
at which we will report allocations at a new high water mark.

We print the list of sampled allocations at the high water mark
whenever we increase that high water mark, on SIGHUP, and on exit.

For short-lived processes, this method may be most useful in
conjunction with `perf record -e sdt_libpoireau:*` and piping the
output of `perf script` to `poireau.py`.

TESTED=manually.